封装axios

+++ doraemon-nodejs/controllers/usersController.js
@@ -124,19 +124,20 @@ const loginUser = (req, res, next) => {
 // 进入用户信息
 const goToUserInfo = (req, res, next) => {
   const token = req.headers.authorization;
-  jwt.verify(token, jwtKey.key, (err, decoded) => {
-    if (err) {
-      return res.send({ state: 2, message: "token 已过期或无效" });
-    }
-    /* const { username } = decoded;
-        usersService.goToUserInfo(username, (err, result) => {
-          if (err) {
-            return res.send({ state: 1, message: err });
-          }
-          return res.send({ state: 0, message: "查询成功", data: result });
-        }); */
-    return res.send({ state: 0, message: "token 验证成功" });
-  });
+
+  const authHeader = req.headers.authorization;
+
+  if (authHeader) {
+    const token = authHeader.split(" ")[1];
+    jwt.verify(token, jwtKey.key, (err) => {
+      if (err) {
+        return res.sendStatus(403);
+      }
+      return res.send({ state: 0, message: "token 验证成功" });
+    });
+  } else {
+    res.sendStatus(401);
+  }
 };
 
 // 收藏
@@ -145,99 +146,179 @@ const doCollect = (req, res, next) => {
 
   const create_time = dateFunction();
 
-  usersService.doCollect(blog_id, user_id, create_time, (err, result) => {
-    if (err) {
-      return res.send({ state: 1, message: err });
-    }
-    return res.send({ state: 0, message: "收藏成功", data: result });
-  });
+  const authHeader = req.headers.authorization;
+
+  if (authHeader) {
+    const token = authHeader.split(" ")[1];
+    jwt.verify(token, jwtKey.key, (err) => {
+      if (err) {
+        return res.sendStatus(403);
+      }
+
+      usersService.doCollect(blog_id, user_id, create_time, (err, result) => {
+        if (err) {
+          return res.send({ state: 1, message: err });
+        }
+        return res.send({ state: 0, message: "收藏成功", data: result });
+      });
+    });
+  } else {
+    res.sendStatus(401);
+  }
 };
 
 // 获取用户信息
 const getUserInfo = (req, res, next) => {
   const { id } = req.body;
-  usersService.getUserInfo(id, (err, result) => {
-    if (err) {
-      return res.send({ state: 1, message: err });
-    }
-    return res.send({ state: 0, message: "查询成功", data: result });
-  });
+
+  const authHeader = req.headers.authorization;
+
+  if (authHeader) {
+    const token = authHeader.split(" ")[1];
+    jwt.verify(token, jwtKey.key, (err) => {
+      if (err) {
+        return res.sendStatus(403);
+      }
+      usersService.getUserInfo(id, (err, result) => {
+        if (err) {
+          return res.send({ state: 1, message: err });
+        }
+        return res.send({ state: 0, message: "查询成功", data: result });
+      });
+    });
+  } else {
+    res.sendStatus(401);
+  }
 };
 
 // 获取用户博客列表
 const getUserBlogList = (req, res, next) => {
   const { id } = req.body;
-  usersService.getUserBlogList(id, (err, result) => {
-    if (err) {
-      return res.send({ state: 1, message: err });
-    }
-    return res.send({ state: 0, message: "查询成功", data: result });
-  });
+  const authHeader = req.headers.authorization;
+
+  if (authHeader) {
+    const token = authHeader.split(" ")[1];
+    jwt.verify(token, jwtKey.key, (err) => {
+      if (err) {
+        return res.sendStatus(403);
+      }
+      usersService.getUserBlogList(id, (err, result) => {
+        if (err) {
+          return res.send({ state: 1, message: err });
+        }
+        return res.send({ state: 0, message: "查询成功", data: result });
+      });
+    });
+  } else {
+    res.sendStatus(401);
+  }
 };
 
 // 获取用户收藏列表
 const getCollectionList = (req, res, next) => {
   const { user_id } = req.body;
-  usersService.getCollectionList(user_id, (err, result) => {
-    if (err) {
-      return res.send({ state: 1, message: err });
-    }
-    return res.send({ state: 0, message: "查询成功", data: result });
-  });
+
+  const authHeader = req.headers.authorization;
+
+  if (authHeader) {
+    const token = authHeader.split(" ")[1];
+    jwt.verify(token, jwtKey.key, (err) => {
+      if (err) {
+        return res.sendStatus(403);
+      }
+      usersService.getCollectionList(user_id, (err, result) => {
+        if (err) {
+          return res.send({ state: 1, message: err });
+        }
+        return res.send({ state: 0, message: "查询成功", data: result });
+      });
+    });
+  } else {
+    res.sendStatus(401);
+  }
 };
 
 // 修改用户头像
 const updateAvatar = (req, res, next) => {
-  //const token = req.headers.authorization;
-  //jwt.verify(token, jwtKey.key, (err, decoded) => {
-  //if (err) {
-  //   return res.send({ state: 2, message: "token 已过期或无效" });
-  // }
-
   const { id, avatarUrl } = req.body;
 
-  console.log(id, avatarUrl);
-  usersService.updateAvatar(id, avatarUrl, (err, result) => {
-    if (err) {
-      return res.send({ state: 1, message: err });
-    }
-    return res.send({ state: 0, message: "修改成功", data: result });
-    // });
-  });
+  const authHeader = req.headers.authorization;
+
+  if (authHeader) {
+    const token = authHeader.split(" ")[1];
+    jwt.verify(token, jwtKey.key, (err) => {
+      if (err) {
+        return res.sendStatus(403);
+      }
+      usersService.updateAvatar(id, avatarUrl, (err, result) => {
+        if (err) {
+          return res.send({ state: 1, message: err });
+        }
+        return res.send({ state: 0, message: "修改成功", data: result });
+      });
+    });
+  } else {
+    res.sendStatus(401);
+  }
 };
 
 // 修改用户信息
 const updateUserInfo = (req, res, next) => {
   const { id, username, nickname, avatarUrl, phone, email, gender, birthday } =
     req.body;
-  usersService.updateUserInfo(
-    id,
-    username,
-    nickname,
-    avatarUrl,
-    phone,
-    email,
-    gender,
-    dateFunction(birthday),
-    (err, result) => {
+  const authHeader = req.headers.authorization;
+
+  if (authHeader) {
+    const token = authHeader.split(" ")[1];
+    jwt.verify(token, jwtKey.key, (err) => {
       if (err) {
-        return res.send({ state: 1, message: err });
+        return res.sendStatus(403);
       }
-      return res.send({ state: 0, message: "修改成功", data: result });
-    }
-  );
+      usersService.updateUserInfo(
+        id,
+        username,
+        nickname,
+        avatarUrl,
+        phone,
+        email,
+        gender,
+        dateFunction(birthday),
+        (err, result) => {
+          if (err) {
+            return res.send({ state: 1, message: err });
+          }
+          return res.send({ state: 0, message: "修改成功", data: result });
+        }
+      );
+    });
+  } else {
+    res.sendStatus(401);
+  }
 };
 
 // 修改密码、手机号、邮箱
 const updateAccount = (req, res, next) => {
   const { id, password, phone, email } = req.body;
 
-  usersService.updateAccount(id, password, phone, email, (err, result) => {
-    if (err) {
-      return res.send({ state: 1, message: err });
-    }
-    return res.send({ state: 0, message: "修改成功", data: result });
-  });
+  const authHeader = req.headers.authorization;
+
+  if (authHeader) {
+    const token = authHeader.split(" ")[1];
+    jwt.verify(token, jwtKey.key, (err) => {
+      if (err) {
+        return res.sendStatus(403);
+      }
+
+      usersService.updateAccount(id, password, phone, email, (err, result) => {
+        if (err) {
+          return res.send({ state: 1, message: err });
+        }
+        return res.send({ state: 0, message: "修改成功", data: result });
+      });
+    });
+  } else {
+    res.sendStatus(401);
+  }
 };
 
 // 用户反馈
@@ -249,34 +330,77 @@ const feedback = (req, res, next) => {
   const snowFlakeId = new SnowFlakeId({ WorkerId: 1 });
   const id = snowFlakeId.NextId();
 
-  usersService.feedback(id, content, user_id, create_time, (err, result) => {
-    if (err) {
-      return res.send({ state: 1, message: err });
-    }
-    return res.send({ state: 0, message: "反馈成功", data: result });
-  });
+  const authHeader = req.headers.authorization;
+
+  if (authHeader) {
+    const token = authHeader.split(" ")[1];
+    jwt.verify(token, jwtKey.key, (err) => {
+      if (err) {
+        return res.sendStatus(403);
+      }
+
+      usersService.feedback(
+        id,
+        content,
+        user_id,
+        create_time,
+        (err, result) => {
+          if (err) {
+            return res.send({ state: 1, message: err });
+          }
+          return res.send({ state: 0, message: "反馈成功", data: result });
+        }
+      );
+    });
+  } else {
+    res.sendStatus(401);
+  }
 };
 
 // 获取用户反馈列表
 const getFeedbackList = (req, res, next) => {
   const { user_id } = req.body;
-  usersService.getFeedbackList(user_id, (err, result) => {
-    if (err) {
-      return res.send({ state: 1, message: err });
-    }
-    return res.send({ state: 0, message: "查询成功", data: result });
-  });
+  const authHeader = req.headers.authorization;
+
+  if (authHeader) {
+    const token = authHeader.split(" ")[1];
+    jwt.verify(token, jwtKey.key, (err) => {
+      if (err) {
+        return res.sendStatus(403);
+      }
+      usersService.getFeedbackList(user_id, (err, result) => {
+        if (err) {
+          return res.send({ state: 1, message: err });
+        }
+        return res.send({ state: 0, message: "查询成功", data: result });
+      });
+    });
+  } else {
+    res.sendStatus(401);
+  }
 };
 
 // 删除用户反馈
 const deleteFeedback = (req, res, next) => {
   const { id } = req.body;
-  usersService.deleteFeedback(id, (err, result) => {
-    if (err) {
-      return res.send({ state: 1, message: err });
-    }
-    return res.send({ state: 0, message: "删除成功", data: result });
-  });
+  const authHeader = req.headers.authorization;
+
+  if (authHeader) {
+    const token = authHeader.split(" ")[1];
+    jwt.verify(token, jwtKey.key, (err) => {
+      if (err) {
+        return res.sendStatus(403);
+      }
+      usersService.deleteFeedback(id, (err, result) => {
+        if (err) {
+          return res.send({ state: 1, message: err });
+        }
+        return res.send({ state: 0, message: "删除成功", data: result });
+      });
+    });
+  } else {
+    res.sendStatus(401);
+  }
 };
 
 // 管理员获取用户列表
